[chore #155678546]separated business test and authentication test
